requirements.txt
openai==0.27.2
pinecone-client==2.2.1
sentence-transformers==2.2.2
numpy
streamlit==1.25.0
PyPDF2==3.0.0
faiss-cpu